ignore __MACOS files

+++ server.cpp
@@ -208,22 +208,26 @@ static int validate_faust(connection_info_struct* con_info)
             fs::path current_file = fs::path(archive_entry_pathname(my_entry));
             if (gVerbosity >= 1) std::cerr << "archive_read_next_header : " << current_file << std::endl;
 
-            if (current_file.string().substr(current_file.string().find_last_of(".") + 1) == "dsp") {
-                if (!dsp_file.empty()) {
-                    archive_status = archive_read_free(my_archive);
-                    fs::remove_all(tmpdir);
-                    con_info->answerstring = completebutmorethanoneDSPfile;
-                    if (gVerbosity >= 1)
-                        std::cerr << "ERROR, we have more than one dsp file " << current_file << std::endl;
-                    return 1;
+            if (current_file.string().substr(0, 8) == "__MACOSX") {
+                if (gVerbosity >= 1) std::cerr << "Ignore  " << current_file << std::endl;
+            } else {
+                if (current_file.string().substr(current_file.string().find_last_of(".") + 1) == "dsp") {
+                    if (!dsp_file.empty()) {
+                        archive_status = archive_read_free(my_archive);
+                        fs::remove_all(tmpdir);
+                        con_info->answerstring = completebutmorethanoneDSPfile;
+                        if (gVerbosity >= 1)
+                            std::cerr << "ERROR, we have more than one dsp file " << current_file << std::endl;
+                        return 1;
+                    }
+                    dsp_file = current_file.string();
+                    filename = dsp_file;
                 }
-                dsp_file = current_file.string();
-                filename = dsp_file;
+                unzipedDir     = fs::path(tmpdir / current_file).parent_path();
+                string newpath = fs::path(tmpdir / current_file).string();
+                archive_entry_set_pathname(my_entry, newpath.c_str());
+                archive_read_extract(my_archive, my_entry, ARCHIVE_EXTRACT_PERM);
             }
-            unzipedDir     = fs::path(tmpdir / current_file).parent_path();
-            string newpath = fs::path(tmpdir / current_file).string();
-            archive_entry_set_pathname(my_entry, newpath.c_str());
-            archive_read_extract(my_archive, my_entry, ARCHIVE_EXTRACT_PERM);
         }
         // END read archived files
